Small improvements (#32)
* Add __main__.py

Ensure pyprefab can be run as "python -m prefab", and make that
behavior consistent with the script entry point.

* Specify encoding when reading files in text mode

* Update CHANGELOG aesthetics

* Do not create package if target directory already has files

Previously, pyprefab would prompt if the target directory already
had data, but there's too much that could go wrong. Because pyprefab
is intended to initialize a project (rather than update something existing),
it will now report an error when the specified target is not empty.

* Liven up the cli prompts

* Add default values